+++ packages/components/modules/messages/native/__shared__/LeaveGroupDialog/types.ts
@@ -1,0 +1,8 @@
+export interface LeaveGroupDialogProps {
+  open: boolean
+  onClose: VoidFunction
+  profileId: string
+  removingParticipantId: string
+  roomId?: string
+  isSoleAdmin?: boolean
+}