+++ packages/components/modules/navigations/web/NavMini/types.ts
@@ -1,6 +1,8 @@
 import { PartialLogoProps } from '@baseapp-frontend/design-system/components/web/logos'
 
-import { NavigationData } from '../types'
+import { DrawerProps, IconButtonProps } from '@mui/material'
+
+import { NavigationData, SlotProps } from '../types'
 
 export interface NavMiniProps {
   navData: NavigationData
@@ -9,4 +11,7 @@ export interface NavMiniProps {
   LogoIcon?: React.FC
   LogoProps?: PartialLogoProps
   hideToggleButton?: boolean
+  slotProps?: SlotProps
+  VerticalDrawerProps?: Partial<DrawerProps>
+  NavToggleButtonProps?: Partial<IconButtonProps>
 }