+++ packages/components/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @baseapp-frontend/components
 
+## 1.5.0
+
+### Minor Changes
+
+- Navigation components now support extended customization through new optional properties for drawer configuration, slot customization, and toggle button styling, enabling more flexible navigation component theming and behavior configuration.
+## 1.4.20
+
+### Patch Changes
+
+- Added Group Chat creation and Groups Tab
+
+## 1.4.19
+
+### Patch Changes
+
+- Add basic members list
+
+## 1.4.18
+
+### Patch Changes
+
+- Add group details to RN
+- Add Leave group feature
+- Add Mark as unread
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.1.9
+
 ## 1.4.17
 
 ### Patch Changes