+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @baseapp-frontend/design-system
 
+## 1.2.0
+
+### Patch Changes
+
+- Added Group Chat creation and Groups Tab
+
+## 1.1.9
+
+### Patch Changes
+
+- Add Circled Avatar for RN
+
 ## 1.1.8
 
 ### Minor Changes